Move MultivariateNormal to pyro.distributions.torch module
