+++ pyro/distributions/__init__.py
@@ -20,7 +20,6 @@ from pyro.distributions.delta import Delta
 from pyro.distributions.distribution import Distribution  # noqa: F401
 from pyro.distributions.half_cauchy import HalfCauchy
 from pyro.distributions.log_normal import LogNormal
-from pyro.distributions.multivariate_normal import MultivariateNormal
 from pyro.distributions.poisson import Poisson
 from pyro.distributions.random_primitive import RandomPrimitive
 
@@ -33,6 +32,7 @@ from pyro.distributions.torch.dirichlet import Dirichlet
 from pyro.distributions.torch.exponential import Exponential
 from pyro.distributions.torch.gamma import Gamma
 from pyro.distributions.torch.multinomial import Multinomial
+from pyro.distributions.torch.multivariate_normal import MultivariateNormal
 from pyro.distributions.torch.normal import Normal
 from pyro.distributions.torch.one_hot_categorical import OneHotCategorical
 from pyro.distributions.torch.uniform import Uniform
